add user and application policy

+++ app/Http/Controllers/ApplicationController.php
@@ -9,6 +9,7 @@ use App\Models\Attachment;
 use App\Models\User;
 use App\Repositories\ApplicationRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class ApplicationController extends Controller
@@ -34,6 +35,7 @@ class ApplicationController extends Controller
      */
     public function create(Request $request)
     {
+        Gate::authorize('create', Application::class);
         $userId = $request->query('user_id');
 
         $user = User::findOrFail($userId);
@@ -51,6 +53,7 @@ class ApplicationController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Application::class);
         $request->validate([
             'user_id' => 'required|exists:users,id',
             'category' => 'required',
@@ -97,6 +100,7 @@ class ApplicationController extends Controller
      */
     public function edit(Application $application)
     {
+        Gate::authorize('update', $application);
         return view('applications.form', [
             'application' => $application,
             'categories'  => ApplicationCategory::cases(),
@@ -110,6 +114,7 @@ class ApplicationController extends Controller
      */
     public function update(Request $request, Application $application)
     {
+        Gate::authorize('update', $application);
         $request->validate([
             'category' => ['required', new \Illuminate\Validation\Rules\Enum(ApplicationCategory::class)],
             'status'   => ['required', new \Illuminate\Validation\Rules\Enum(ApplicationStatus::class)],
@@ -129,6 +134,7 @@ class ApplicationController extends Controller
      */
     public function destroy(Application $application)
     {
+        Gate::authorize('delete', $application);
         foreach ($application->attachments as $attachment) {
             Storage::disk('public')->delete($attachment->file_path);
         }
